[build] [refactor] Move shaders source files to runtime (#5247)
* Move shaders to runtime

* Update path

* Add dep for gl
