+++ taichi/runtime/metal/shaders/ad_stack.metal.h
