+++ taichi/runtime/metal/shaders/init_randseeds.metal.h
